fix: remove UB when changing state (#865)

+++ src/raft/recv.c
@@ -72,7 +72,7 @@ static int recvMessage(struct raft *r, struct raft_message *message)
 	};
 
 	if (rv != 0 && rv != RAFT_NOCONNECTION) {
-		tracef("recv: %d: %s", message->type, raft_strerror(rv));
+		tracef("recv: %d: %s", (int)message->type, raft_strerror(rv));
 		return rv;
 	}
